feat!: Use standard library version of Option and Result (#371)
Adds `dafny-lang/libraries` as a submodule, and replaces `Option` and `Result` in `src/StandardLibrary/StandardLibrary.dfy` with its version.

The changes to the rest of the files are search-and-replace changes necessary to account for:

1. The new `Option` using `value` instead of `get`, and `UnwrapOr` instead of `GetOrElse` (which has a much less threatening tone I have to say :); and
2. The new `Result` having two type parameters rather than one (the old version was hardcoded to use a `string` for the error value).

In many places `import opened StandardLibrary` was ONLY necessary to pull in these super common wrapper types, so I replaced it with `import opened Wrappers`. I expect as more standard library code is dropped or moved into the shared libraries repo itself, the `StandardLibrary` module will be eventually dropped in favour of more, smaller modules in the style of `dafny-lang/libraries`.