refactor(NET): NativeWrappers validate output
Validating Output revealed an issue with Excpetions being cast
from Base to the concrete type.

Resolving that exposed that our exception messages from NativeWrapper
were not very helpful... so I fixed that by refactoring the validation
call into nested method so the ArguementException could be re-cast
as a Service Exception but with a helpful message.