fix(java:gen:awssdk:ToNative): Collections of enums must be arrays
In the AWS SDK for Java V1,
collections of enums MUST be input as either:
1. A collection of strings
2. An array of enums

At this time,
I have opted for an array of enums.
This allows us to continue generating converters
that use List or Sets of Enums.

It is not performant:
Our generated list of enums will be
copied to a new array,
and the AWS SDK will then iterate over that array,
converting its members to strings,
and storing that in a new list.

But, to my knowledge, we never call any AWS SDK APIs that take lists of enums.
And, this is an issue ONLY for the AWS SDK Java V1...