feat: Add wraping for dotnet
Adding wrapping.
But also needed to adjust the visibility of a few things.

So that a dependent module can use and reason about many of these generated components
the correctness elements MUST be public.

This mostly includes `isSet` checks.
But also the TypeConversion classes.
For internal conversions these remain `internal`.
But the class is now public,
as are conversions for reference types.

This is because a dependent module
will not have access to the wrapping classes.
But it sill needed to be able to pass
the element across the Dafny boundary.
So conversions of this type are not Public.

And in dependent TypeConversion
these public methods will be used
to convert such elements across the Dafny boundary.