feat(Java): Half-Baked native wrapper
I am honestly not sure what work remains.
But Ryan strongly encourages me to put this down
and work on something else...
Even though I really want to finish this.

When I get back to this,
I should revert or undo this commit,
and break it into smaller pieces.
There also maybe work related
to this in my local stash
or IntelliJ shelf.