feat(Polymorph-CLI)!: refactor CLI arguements
BREAKING CHANGE:
`--output-dafny` now takes a Path,
just like `--output-dotnet` and `--output-java`.

`--output-local-service-test` is now
`--local-service-test` and does NOT take path.

Invoking `--output-local-service-test` will cause an error.
Invoking `--local-service-test` with a PATH will cause an error.

However, to mitigate transition pain,
if `--ouput-dafny` is given with no argument,
the generated Dafny will be placed in the Model path.

Otherwise, Generated Dafny will be written to the value of
`--output-dafny`.

Closes https://github.com/awslabs/smithy-dafny/issues/104