chore: Avoid direct references to java conversion library in codegen (#210)
The code generation libraries should not directly depend on libraries just because the code they generate might. Although it offers a slight convenience in ensuring the symbols are correct, it bloats the codegen dependencies more and more over time, and can lead to false version conflicts between what the codgen needs and what the code it generates needs.

This PR just breaks the dependency on software.amazon.dafny:conversion, but we should break the dependency on dafny.lang:DafnyRuntime as well in the future.