try fixing CI
`gradle build` in the SQS test model directory, works on macOS but not
on Linux (in CI). hypothesis: this is because the Smithy build tooling
automatically searches for a "model" directory adjacent to the
smithy-build.json, which passes on macOS only due to the
case-insensitive-by-default filesystem.

this commit tries to work around that by explicitly specifying the
"Model" directory as a model source directory.

see: https://smithy.io/2.0/guides/building-models/build-config.html#using-smithy-build-json