fix: Always pass -compileSuffix:1 (#316)
Fixes the nightly build: https://github.com/smithy-lang/smithy-dafny/actions/runs/7818075629/job/21327443046

Dafny master stopped including this option in dafny /help since it's intended to be a hidden option, which breaks the part of the build that checks if the Dafny version supports it. But this project and all known smithy-dafny projects are all already on at least Dafny 4.2 so they can all just pass the option unconditionally now.