chore: Code generation target for SimpleString test model in Rust (#332)
Manually-crafted code generation target for SimpleString in Rust. Still with some TODOs around error handling, but includes a working set of tests showing what it's like to use the smithy-rs-generated interface around the Dafny implementation.

See this README.md for more details. I'm particularly interested in feedback on the conversions module, since that's the most novel part compared to all of the boilerplate generated by the existing smithy-rs code generator.

---------

Co-authored-by: seebees <ryanemer@amazon.com>
Co-authored-by: Mikael Mayer <mimayere@amazon.com>