chore: Change service constructor result type back (from trait to concrete class) (#331)
*Description of changes:*

Reverts the part of https://github.com/smithy-lang/smithy-dafny/pull/301 that changed the result type of constructing services from the concrete class type to the trait type, which turned out to be breaking in .NET as well as Dafny and was manually reverted in https://github.com/aws/aws-cryptographic-material-providers-library/pull/257. See https://github.com/aws/aws-cryptographic-material-providers-library/pull/267 for the upcoming proper fix to the MPL using this change.

As it turns out this change wasn't necessary in the first place, and the `CreateSuccessOfClient/CreateFailureOfError` helpers are unchanged and still use the trait type, because these helpers are only needed for services that are created by target language code, i.e. SDK-style services and `--local-service-test` wrapped services. Regular local services are constructed in Dafny code so they never need the helper methods.