chore: kms-lite Rust beneration (#473)
Issue #, if available:
#413

Description of changes:
Fills MOST of the code generation target for kms-lite. Even with the operations trimmed down to only those used by the MPL, this is still a ton of boilerplate, so I mostly only implemented the operations that are actually tested (Encrypt, Decrypt, and GenerateDataKey). I may at least partially implement the beginnings of codegen to come back and fill in the rest as well as tackle #477.