fix: Update Rust runtime in nightly build, fix regression from Rust runtime change (#527)
The Dafny runtime library made a (reasonable) breaking change on `feat-rust`, and the nightly build didn't catch it because it doesn't use the matching runtime. Fixed the nightly build to do so and confirmed it exposes the problem (https://github.com/smithy-lang/smithy-dafny/actions/runs/10493792057/job/29068594398#step:10:915), and then updated the `feat-rust` commit used in CI and fixed the various copies of `standard_library_conversions.rs` and a few straggling direct references accordingly.

Also spotted that the SDK codegen had the path to `standard_library_conversions.rs` wrong and hence wasn't actually getting applied over what's checked in for `ddb-lite` and `kms-lite`.