chore(ci): make all pull.yml jobs required in branch protection automatically (#554)
As is the case for many suites of GitHub actions, it's a huge pain to manually maintain the list of required checks in the branch protection. This change leverages https://github.com/re-actors/alls-green to add a single capstone action that only passes if all dependent jobs pass (see repo README for details on why the naive approach doesn't work).

Once this is approved I'll update the required checks before merging.

Also adding Python tests to the list of triggers in the nightly build since I noticed it was missing.