feat: Better .NET type conversion encapsulation
Dafny has optional parameters.

Smithy-Dafny uses this feature
to produce Dafny types
that default to `:= None`.

However, the translated Dafny
does *not* support optional parameters.

This means that these optional parameters,
are _not_ interoperable in target runtimes.

.NET already deferred type conversions
of wrapped resources to the target module.
But it assumed that Dafny was overly interoperable.

This change to .NET defers all type conversion
to the module that owns the type.
This brings .NET more in line with other runtimes.

This change is not completely backwards compatible.
There are a few edge cases with orphaned types
as well as already published versions.