fix(go): utf8-utf16 encoding (#764)
*Issue #, if available:*

*Description of changes:*
This PR fixes the utf8-utf16 encoding issue. In Go, strings are utf8
encoded (but underlying byte can be anything). With `unicode-char:false`
in dafny, strings are utf16 encoded. This means anything going over from
Go to Dafny needs to go through utf8-to-utf16 encoding and vice versa.

The issue was that dafny uses a `Char` type which is an int32 / rune to
encode utf16 bytes, and not the uint16. This causes utf16 bytes to pass
through as utf8 bytes in Go. But then with surrogate pairs, it all falls
apart because dafny uses two int32 bytes to encode two surrogate pairs
which can (and should be) fit in 2 uint16 / 1 int32 .

Read:
https://github.com/dafny-lang/dafny/blob/master/Source/DafnyCore/Backends/GoLang/GoCodeGenerator.cs#L2355-L2377

Anything annotated with dafnyutf8bytes in smithy needs to pass through
as it is between the two layers.

This PR also fixes a missed validation with utf8bytes and does some
optimization for type conversion layer.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

# Check any applicable:
- [ ] Were any files moved? Moving files changes their URL, which breaks
all hyperlinks to the files.

---------

Co-authored-by: Shubham Chaturvedi <scchatur@amazon.com>